[MY_Controller] Error in initlialiting
Putting database in "load->library" will cause an error when reacing the
base url.

correct format is to load esternally the helper.

p.s. if you dont insert the "$this->load->database()" part, and even not
loading it by the "load->library", seems to not cause any problem to the
application.