Bug Fixes
Found and fixed a few minor bugs in the code.

On the users profile page, the achievements.php library is not loaded
for users that are logged out,  this has now been changed.

Page title in the browser was not showing for the users profile page.

Page title was not showing up on a full reply to a discussion.