Huge Update
This is a huge update,  I have restructured the database to make it
more efficient and to stop relying on Join statements to do all the
work.

Have also implemented MY_Model.php to handle the database work and to
keep the code DRY.

Have updated the installer to reflect all the database changes.

Many variable names have been updated and the view updated to reflect
this.

I am still updating a lot of the features so their will be broken code
for the time been.